+++ DiveDeepWebApp/DiveDeepWebApp/ViewModels/WeatherViewModel.cs
@@ -14,7 +14,7 @@ namespace DiveDeepWebApp.ViewModels
 
         public double WindSpeed10m { get; set; } //always km/h - if < 8 m/s → all good, if > don't swim
 
-        public int Precipitation { get; set; } // always mm - if < 2mm/t → all good, if > don't swim
+        public double Precipitation { get; set; } // always mm - if < 2mm/t → all good, if > don't swim
 
         public bool Thunder { get; set; } //if 95-99 or 17 or 29 = thunder, don't swim
